Backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
pymongo[asyncio]
beanie>=1.25.0
python-jose[cryptography]
passlib[bcrypt]
pyjwt
python-multipart